extract instrumentation processor into separate library
