+++ sintr_analysis_server_example/lib/session_metrics.dart
